refactor(samples): split drag & drop into sortable and kanban demos
The /drag-drop showcase hosted both demos in one 178-line DragDropDemo
rendered through a single CodeSample. Extract them into DragDropSortableDemo
and DragDropKanbanDemo, each with its own scoped CSS, shown as separate
CodeSample cards under H2 headers — matching the multi-demo layout already
used by /virtualize and /binding. Same route, test-ids, and behaviour.
