+++ samples/Rask.Example.Shared/Features/DragDrop/DragDropKanbanDemo.cs
@@ -2,9 +2,9 @@ using Rask.Core.DragAndDrop;
 
 namespace Rask.Example.Shared.Features;
 
-public sealed class DragDropDemo : Component
+// Multi-column board. One drop zone per column; drag cards across columns or reorder within one.
+public sealed class DragDropKanbanDemo : Component
 {
-    // Demo B — Kanban board. One drop zone per column.
     private static readonly string[] _columns = ["todo", "doing", "done"];
 
     private static readonly Dictionary<string, string> _columnLabels = new()
@@ -22,82 +22,7 @@ public sealed class DragDropDemo : Component
         ["done"] = [new Card(5, "Read the codebase")]
     };
 
-    // Demo A — single-list reorder. One drop zone ("list").
-    private readonly List<string> _fruits =
-    [
-        "Apple", "Banana", "Cherry", "Date", "Elderberry"
-    ];
-
-    protected override RenderResult Render() =>
-    [
-        SortableDemo(),
-        KanbanDemo()
-    ];
-
-    // ----- Demo A: sortable list ---------------------------------------------
-
-    private Component SortableDemo() =>
-        Div(Class: "card shadow-sm border-0 mb-4")[
-            Div(Class: "card-body")[
-                H5(Class: "fw-semibold mb-1")["Sortable list"],
-                P(Class: "text-secondary small")["Drag a fruit and drop it onto another to reorder. One drop zone."],
-                DragDrop(
-                    SortableBody,
-                    ReorderFruit)
-            ]
-        ];
-
-    private Component SortableBody(DragDropContext ctx)
-    {
-        var rows = new List<Child>(_fruits.Count);
-        for (var i = 0; i < _fruits.Count; i++)
-        {
-            var fruit = _fruits[i];
-            var index = i;
-            var cls = "list-group-item d-flex align-items-center gap-2 dd-item";
-            if (ctx.IsSource("list", index))
-            {
-                cls += " dd-dragging";
-            }
-
-            if (ctx.IsDropTarget("list", index))
-            {
-                cls += " dd-drop-target";
-            }
-
-            rows.Add(Li(
-                Key: fruit,
-                Class: cls,
-                Draggable: true,
-                OnDragStart: ctx.DragStart("list", index),
-                OnDragOver: ctx.DragOver("list", index),
-                OnDrop: ctx.Drop("list", index),
-                OnDragEnd: ctx.DragEnd,
-                Data: new Dictionary<string, string?> { ["testid"] = $"fruit-{index}" })[
-                I(Class: "bi bi-grip-vertical text-secondary"),
-                Span(Class: "fw-semibold")[fruit]
-            ]);
-        }
-
-        return Ul(Class: "list-group dd-list", Id: "dd-fruit-list")[rows];
-    }
-
-    // Direction-aware: dragging down lands after the target, dragging up lands before it.
-    private void ReorderFruit(DragDropMove move) => move.ApplyTo(_fruits);
-
-    // ----- Demo B: Kanban board ----------------------------------------------
-
-    private Component KanbanDemo() =>
-        Div(Class: "card shadow-sm border-0 mb-4")[
-            Div(Class: "card-body")[
-                H5(Class: "fw-semibold mb-1")["Kanban board"],
-                P(Class: "text-secondary small")[
-                    "Drag cards between columns, or reorder within one. Each column is a drop zone."],
-                DragDrop(
-                    KanbanBody,
-                    MoveCard)
-            ]
-        ];
+    protected override RenderResult Render() => DragDrop(KanbanBody, MoveCard);
 
     private Component KanbanBody(DragDropContext ctx)
     {
