+++ samples/Rask.Example.Shared/Features/DragDrop/DragDropKanbanDemo.css
@@ -1,12 +1,3 @@
-.dd-item {
-    cursor: grab;
-    user-select: none;
-}
-
-.dd-item:active {
-    cursor: grabbing;
-}
-
 .dd-dragging {
     opacity: 0.4;
 }
